cleaned code

+++ dataset.py
@@ -7,10 +7,11 @@ from torchtext import datasets
 
 def load_data():
     if not os.path.exists('./data/snli_1.0'):
-        print('Downloading')
-        r = requests.get('https://nlp.stanford.edu/projects/snli/snli_1.0.zip', allow_redirects=True)
-        open('./data/snli_1.0.zip', 'wb').write(r.content)
-        os.system('unzip ./data/snli_1.0.zip')
+        if not os.path.exists('./data/snli_1.0.zip'):
+            print('Downloading')
+            r = requests.get('https://nlp.stanford.edu/projects/snli/snli_1.0.zip', allow_redirects=True)
+            open('./data/snli_1.0.zip', 'wb').write(r.content)
+        os.system('unzip ./data/snli_1.0.zip -d ./data/')
 
     with open('./data/snli_1.0/snli_1.0_train.jsonl') as f:
         train = np.array(list(map(lambda x: {k:v for k, v in json.loads(x).items() if k in ['sentence1', 'sentence2', 'gold_label']}, f.readlines())))
